Make test visible for Django>=1.6 testrunner.
